Update route to include form so it is clear what step refers to Move submit route into the same file and the other form steps
